feat: unwrap auth namespace to flat exports + barrel
